requirements.txt
# Python dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
boto3>=1.28.0
pyyaml>=6.0
kafka-python>=2.0.2
streamlit>=1.28.0
